Move reduce2d into reduce folder
